pyproject.toml
[tool.poetry]
name = "programlamadilleri_api"
version = "0.1.0"
description = ""
authors = ["ahmetveburak <ahmetbozyurtt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
psycopg2 = "^2.8.6"
django-multiselectfield = "^0.1.12"
djangorestframework = "^3.12.4"
djangorestframework-api-key = "^2.0.0"
Pillow = "^8.3.1"
easy-thumbnails = "^2.7.1"
django-filter = "^2.4.0"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
mypy = "^0.900"
isort = "^5.9.3"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
